Delete obsolete Application.mk
The `Application.mk` file of `LiquidFun` is no longer needed. In fact, its mere existence seems to disturb the build in such way that `orx` crashes intermittently when resuming the game.

One theory is that the `LiquidFun` library inadvertently targets an older platform with no 16 kB page support. All in all, confusing.

Thus, it makes sense to simply delete the file.